Add tts mp3 generator
I was unable to add a live tts generator via the phone.
An alternative was to pre-generate audio on the computer, then google
drive automatically uploads those files. This makes the entire project
somewhat useless to others unless they backup these files with gdrive.
But for me, it works.

Also, there was a weird issue. Writing the file 'con.mp3' results in
an error on windows because 'con' is a reserved word that can't be
used for file names. I renamed the word 'kon' in Spanish.